# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentacion de API de victor andre mejia castellanos",
        "description": "API desarrollada por victor andre mejia, para la Universidad Francisco Gavidia",
        "contact": {
            "email": "ia.andycastellanos@ufg.edu.sv"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        },
        {
            "url": "http://testing.ejemplo.com/",
            "description": "Servidor para pruebas"
        },
        {
            "url": "http://klrm.ejemplo.com",
            "description": "Servidor de produccion"
        }
    ],
    "paths": {
        "/api/zonas": {
            "get": {
                "tags": [
                    "Zona"
                ],
                "summary": "Obtener toda las zonas registradas",
                "description": "Devuelve todas las zonas registradas en la tabla zona de la base de datos, no requiere parámetros",
                "operationId": "obtenerZonas",
                "responses": {
                    "200": {
                        "description": "Zonas devueltas satisfactoriamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Valores encontrados"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_zona": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_pais": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "nombre_zona": {
                                                        "type": "string",
                                                        "example": "Zona Occidental"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "code": {
                                                        "type": "integer",
                                                        "example": 200
                                                    },
                                                    "msg": {
                                                        "type": "string",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": "4"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron valores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "No se han encontrado valores"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": ""
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "code": {
                                                        "type": "integer",
                                                        "example": 404
                                                    },
                                                    "msg": {
                                                        "type": "string",
                                                        "example": "La base de datos no devolvió registros"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}